backend/requirements.txt
# Core Framework
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.31
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# ML & Data Science
scikit-learn==1.7.2
numpy==2.3.3
pandas==2.3.2
joblib==1.5.2
nltk==3.8.1

# AI & LLM Integration
openai==1.35.13
anthropic==0.29.0

# Utilities
python-dotenv==1.1.1
httpx==0.25.2
